pyproject.toml
[tool.poetry]
name = "looptrace"
version = "0.3dev"
description = "Library and programs for tracing chromatin loops from microscopy images"
authors = [
    "Kai Sandvold Beckwith",
    "Vince Reuter <vince.reuter@gmail.com>",
]
readme = "README.md"
include = ["looptrace/looptrace-assembly*.jar"]
license = "MIT"
repository = "https://github.com/gerlichlab/looptrace"
classifiers = [
    "Development Status :: 1 - Planning",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Topic :: Bioinformatics",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
# These are the main runtime dependencies.
python = ">= 3.10, < 3.11"
cellpose = { version = "v2.2.2", optional = true }
czifile = { version = "^2019.7.2", optional = true }
dask = "^2023.5.1"
flowdec = { version = "^1.1.0", optional = true }
gertils = { git = "https://github.com/gerlichlab/gertils.git", tag = "2023-08-01" }
# TODO: imagecodecs is used by nd2 to support "legacy" ND2 (JPEG2000), but may not be needed otherwise.
imagecodecs = { version = "2023.9.18", optional = true }
joblib = { version = "^1.3.1", optional = true }
matplotlib = { version = "^3.7.0", optional = true }
napari = { version = "^0.4.17", optional = true }
napari-plot-profile = { version = "^0.2.0", optional = true }
nd2 = { version = "^0.5.3", optional = true }
numba = { version = "^0.57.0", optional = true }
numcodecs = "^0.11.0"
numpy = ">= 1.24.2, < 1.25"
pandas = "^1.5.3"
piper = { version = "0.12.3", optional = true }
pyqt5 = { version = "5.15.7", optional = true }
pyqt6 = { version = "^6.5.0", optional = true }
pyside6 = { version = "^6.5.0", optional = true }
pytest = { version = "^7.3.1", optional = true }
pytest-skip-slow = { version = "^0.0.5", optional = true }
pyyaml = "^6.0"
scipy = "^1.10.1"
scikit-image = "^0.20.0"
seaborn = { version = "^0.12.0", optional = true }
tensorflow = { version = "2.11.1", optional = true }
tensorrt = { version = "8.6.1", optional = true }
tifffile = { version = "^2023.4.12", optional = true }
tqdm = "^4.65.0"
zarr = "^2.4.12"

[tool.poetry.extras]
analysis = ["matplotlib", "numba", "seaborn"]
deconvolution = ["flowdec"]
dev = ["czifile", "matplotlib", "joblib", "nd2", "numba", "pytest", "pytest-skip-slow", "seaborn", "tifffile"]
interactive = ["napari", "napari-plot-profile", "pyqt6"]
legacy-support = ["imagecodecs"]
nuclei = ["cellpose"]
pipeline = ["cellpose", "czifile", "flowdec", "joblib", "nd2", "piper", "tensorflow", "tensorrt", "tifffile"]
polymer = ["numba"]
test = ["piper", "pytest"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
